+++ lib/model/app/contributors.dart
@@ -27,7 +27,7 @@ class Contributors {
     Contributor('Oğuz Ersen', 'oguz@ersen.moe'),
     Contributor('Allan Nordhøy', 'epost@anotheragency.no'),
     Contributor('pemibe', 'pemibe4634@dmonies.com'),
-    Contributor('Linerly', 'linerly@protonmail.com'),
+    Contributor('Linerly', 'linerly@proton.me'),
     Contributor('Skrripy', 'rozihrash.ya6w7@simplelogin.com'),
     Contributor('vesp', 'vesp@post.cz'),
     Contributor('Dan', 'denqwerta@gmail.com'),
@@ -117,14 +117,17 @@ class Contributors {
     Contributor('Petrov', 'i_v_c@mail.ru'),
     Contributor('தமிழ்நேரம்', 'anishprabu.t@gmail.com'),
     Contributor('bovirus', 'roberto.boriotti@canon.it'),
-    Contributor('Максим Горпиніч', 'mgorpinic2005@gmail.com'),
+    Contributor('Максим Горпиніч', 'maksimgorpinic2005a@gmail.com'),
     Contributor('Priit Jõerüüt', 'hwlate@joeruut.com'),
+    Contributor('Dr Hieu', 'v7pvas7m@anonaddy.com'),
+    Contributor('Saúl Palacios', 'palacios22c@gmail.com'),
+    Contributor('Grooty12', 'Rasmus@rosendahl-kaa.name'),
+    Contributor('Victor M', 'victormorita@tuta.io'),
+    Contributor('cat', 'catsnote@proton.me'),
     // Contributor('Femini', 'nizamismidov4@gmail.com'), // Azerbaijani
     // Contributor('Alvi Khan', 'aveenalvi@gmail.com'), // Bengali
     // Contributor('Htet Oo Hlaing', 'htetoh2006@outlook.com'), // Burmese
     // Contributor('Khant', 'khant@users.noreply.hosted.weblate.org'), // Burmese
-    // Contributor('Grooty12', 'Rasmus@rosendahl-kaa.name'), // Danish
-    // Contributor('Victor M', 'victormorita@tuta.io'), // Danish
     // Contributor('Åzze', 'laitinen.jere222@gmail.com'), // Finnish
     // Contributor('Olli', 'ollinen@ollit.dev'), // Finnish
     // Contributor('Idj', 'joneltmp+goahn@gmail.com'), // Hebrew